always "enable" filter extension for Eden repos
Summary:
This diff marks the filter extension as always enabled. The previous diffs in this stack make the filter extension no-op unless "filteredhg" is written to .hg/requires, so this should have no consequences for normal or sparse clones.

"filteredhg" isn't written to .hg/requires in any existing workflows. The next diff in the stack introduces a filteredfs clone option for `eden clone` that writes to .hg/requires and sets-up the required files for the filtered extension to work.

Reviewed By: muirdm

Differential Revision: D50992737

fbshipit-source-id: 559eacea5e550a52ab7e1d6427039f94edc32123