implement `eden filteredfs enable`
Summary: This diff implements a half-complete version of `eden filter enable`. The only missing part is accurate sparsematcher() construction (right now this is stubbed to the AlwaysMatcher). Sparsematcher() construction will be implemented in the next few diffs. For now, we just check that the enable command writes the correct file to the repo.

Reviewed By: muirdm

Differential Revision: D51325069

fbshipit-source-id: 42f4a1aef80f72973aaa3be277c05710c87a0a83