change expected edensparse config format
Summary:
We previously expected the edensparse file to only contain the file name. To make the config file more compatible with non-eden sparse code, let's expect the file to contain just a single line: `%include {filter_path}`

This will be used in the next diff so we can reuse sparse profile code.

Reviewed By: muirdm

Differential Revision: D51333046

fbshipit-source-id: ef9625584554736030e6b5fe504d1a7db6dad05f