move common attributes to mixin class
Summary:
I want to share a small number of attributes across the SparseRepo and EdenSparseRepo classes, but deriving from the SparseRepo class isn't possible since it's defined in the scope of the _wraprepo() function.

To work around this, I define a SparseMixin class that implements these common attributes that I want to share. The SparseRepo and EdenSparseRepo class both use this mixin to share the common attributes.

These shared attributes are used in the D51325069

Reviewed By: muirdm

Differential Revision: D51333048

fbshipit-source-id: d017886cfbc44fdc15e4ff9e3e3e3d8c3552021f