tests: remove last remnant of QueuedImmediateExecutor
Summary:
Most of it was replaced with a ManualExecutor. Some care had to be taken around
the use of StoredObject as triggering an error on them when no future was
obtained from it leads to the error not being triggered when the future is then
retrieved from it.

The only place where QueuedImmediateExecutor is still present is in the
ImmediateFutureTest to demonstrate the lack of safety with it.

Reviewed By: genevievehelsel

Differential Revision: D51302393

fbshipit-source-id: fb66a9426e0ae7672799bd997898e8b525ff112c