trees: short circuit request for null commit's manifest
Summary:
An empty working copy has no parent commit, which the code sometimes treats as having the null commit as a parent.

Previously when getting the null commit's manifest, we would end up issuing a remote request for the null tree id, which is pointless.

Fix by short circuiting null commit ids in ReadTreeManifest trait.

I'm not sure this is the best place to do this, but it seemed to do the trick.

Reviewed By: mitrandir77

Differential Revision: D51723090

fbshipit-source-id: 9de9201f1a559c799ca8483ee56a2466c432ba8b