goto: add a function for clearing EdenFS treestate in Rust
Summary:
When running `sl goto` the EdenFS treestate needs to be cleared, which essentially means notifying EdenFS that files that were added/modified/removed have returned to their previous state.

In order to achieve this, the `workingcopy::util:walk_treestate` was modified so that it could also match against `any` instead of only `all` and `none`.

Reviewed By: muirdm

Differential Revision: D51482436

fbshipit-source-id: f1f68760090301fa54a4c730a7c65c6f467c777c