mononoke_types: disallow newlines in commit author and committer
Summary:
The commit author field is stored in a newline-separated field in both Mercurial and Git data formats.  This means it is disallowed to appear in either of these.  Currently we don't enforce that in Mononoke, which means we can have bonsai changesets that are not representable in Mercurial or Git.

Disallow newlines in well-formed bonsai changesets also.  This will prevent this situation from occurring.

Differential Revision: D51807680

fbshipit-source-id: b66807a6a22a0ed5ada971206afd9deb3f29088d