previews: sort commits by "preview insertion order" seqNumber
Summary:
In D51784283, I discovered the "preview" branch should be put on the right side
regardless of the commit date. But that's not true for all kinds of previews.

For example, when confirming a "hide" by setting previewType, it could be:

  b [hide: yes | no ?]
  |
  | c
  |/
  a

In this case 'b' should stay at its original place, not swap with 'c'. Note the
same confirmation can happen on 'c' too and 'c' should not move to the left
side of 'b'.

While it's possible to special case the "hide" preview, I think a more robust
solution is to track the actual "insertion" order to the graph. Commits last
inserted (by preview logic) are on the right side.

This is done by keeping an "insertion sequence number". When editing a commit,
the logic can decide whether to keep the seq number (ex. hide) or request a new
one (ex. rebase). Practically, all commits from the "log" output have a same
seqNumber (0), commits edited by previews might have higher numbers.

Reviewed By: evangrayk

Differential Revision: D51813803

fbshipit-source-id: 2fcab3e308b50342943237bd15d43d7d28b95615