dag: add rebase API
Summary:
It turns out that there are multiple places that needs rebase preview logic.
So let's add one.

Rebase can be tricky when it has to leave "orphaned" commits. For example,

    d                              d  <- note the parent of d is neither c nor y
    |                              |
    c  --> rebase -r b+d -d y -->  b c
    |                              | |
  y b                              y b <- old b is still there but obsoleted
  |/                               |/
  x                                x

This rebase implementation handles the above case. It does not handle existing
mutation info (skip rebasing when successor in ancestors(dest); or copy 'x
amended to y' to new 'x' and 'y' if both 'x' and 'y' are being rebased) though.

Reviewed By: evangrayk, zzl0

Differential Revision: D51784277

fbshipit-source-id: 38b80cb5cb6caa23d878fa811a8d28f28dcefc0e