dag: initial class for "set" representation
Summary:
To provide revsetlang-like queries. We need a dedicated "Set" type to
represent a set of commit hashes. This diff adds it. For now it's just
a thin wrapper around `immutable.Set<Hash>`. We might want to add more
information to the set type in the future, like preserving the topology
order (which hg's revset already does), etc. But this is a start.

Reviewed By: muirdm, evangrayk

Differential Revision: D51712900

fbshipit-source-id: 1379886ac98da348c65268b6f26049910ce6ab3c