dag: split out low-level dag implementation
Summary:
The `dag` has 2 parts:
- The basic part, that only needs to know "parents" to work.
- Extended parts, like "obsolete", "public", "successors", etc.

In the future I'd like to make the extended dag also maintain the
mutation/successor relationship. In that case the mutation dag should be just
the basic dag. So let's use 2 types to make it clear.

This removes a bunch of `as Partial<CommitInfo>` and unsound type casts that
will not be accepted by flowtype (although we don't use flow at this point).

Reviewed By: evangrayk

Differential Revision: D51813842

fbshipit-source-id: e5260fcc5443113a41acd4e68854568654877172