previews: use dag provided successors
Summary:
Now `dag` (with previews) is able to answer successor questions, use it.
This allows following successors with previews, which can be useful in stacked
preview cases (ex. split in the middle).

Note: although some faked hashes are used in the preview, they don't leak to
the command line yet, since the command line just uses `successors` revset.
So that will not be an issue.

Reviewed By: evangrayk

Differential Revision: D51813830

fbshipit-source-id: fe54da60e7fbe12e63b7d83160cfb47d8cc2cbfb