repolock: repurpose RepoLockHandle as a dereffable &Path
Summary:
This way a working copy lock can be passed into other library code to not only act as proof of lock acquisition, but to also act as the working copy ".hg" path.

This will be useful in the repostate crate which will be reading the various ".hg" state files (and wants to be certain the working copy lock is held).

Reviewed By: quark-zju

Differential Revision: D51771502

fbshipit-source-id: 12b9cf1c92749fa0080fdcf683ede726e1cd9912