Propagate errors from backingstore up through stack
Summary:
SaplingNativeBackingStore "eats" errors for single object fetches. This prevents certain kinds of logging to scuba and other actions that might be taken by higher layers of code.

Eventually, we wil changes these APIs to return futures, but in the meantime, we will switch them to return folly::Try's.

The propagation goes from SaplingNativeBackingStore to HgDatapackStore to HgBackingStore and all of their callers.

Reviewed By: kmancini

Differential Revision: D51777559

fbshipit-source-id: 177e97fc12a5438abe0ba83f97aa3b1367074b3a