Migrate forward syncer to new clap and MononokeApp
Summary:
As title, moving the forward syncer binary to use the newer clap and Mononoke API, which is the newer (and better) way to write CLIs that work with Mononoke.

I usually split diffs like this, but in this case it would likely be very tricky to do so, unless maybe I created a new CLI temporarily and moved the commands one by one.
I was going to do that but pierrec mentioned he would prefer to review everything in a single diff, so here it is 😅

Other minor changes:
- I enabled logging to the `mononoke_x_repo_sync` Scuba table by default. Doesn't look like we're doing that.
- Added doc comments to some commands, args and functions.

## Notes & FYI
- I'm adding doc comments to the commands/arguments that I have a high degree of certainty are correct. The tool currently has no helpers/doc comments, by not adding to all commands/args at least I won't be making it worse 😅

## RFC:
- Should I use `create_commit_syncers_from_app_unredacted` or `create_commit_syncers_from_app`?
- Please double check the doc comments I added to the arguments and the functions. They're my effort to provide some documentation, rather than none.

Reviewed By: mitrandir77

Differential Revision: D51583743

fbshipit-source-id: 840bdd8275fd607c66f7280c4a4443e0abdd7254