pyedenapi: change pushvars's type from Vec to PyDict
Summary:
the underlying edenpi APIs expect HashMap<String, String>, it would be good
to keep it consistent in the pyedenapi APIs.

Another advantage of PyDict type is that we can use dict in Python, and it's
easier to check a key's value in a dict object than in a list.

Reviewed By: quark-zju

Differential Revision: D51675345

fbshipit-source-id: 6fa4d8c5d026e87d6bba76990b890090b52a239c