commit_graph: prefetch skip-tree edges when processing disjoint segments
Summary:
During processing of disjoint segments, we lower the members of the segment using the skip tree edges.   We can prefetch here for a small (~10%) performance gain when cache cold.

While testing I noticed the integration tests weren't prefetching at all because the JK was defaulting to maximum step count of 0.  I've set the JK value to 4 so that we do *some* prefetching in the tests, but not so much as to prefetch everything in the usually small test repos.

Differential Revision: D51716874

fbshipit-source-id: 5b9412f6d25e90422837c2c9d67bf4754a6560d9