Add UI zoom setting
Summary:
I previously added a setting like this for UI font size. This did not work well, because things are not designed for the font size to vary. It made lots of things look slightly wrong, such as diff badges etc.

Instead, the approach here is to add a zoom setting, so the entire UI scales up together. That way we know that things will be designed as we intended.

Of course, there's always catches, and here the issue is that some width calculations are now different. We need to account for zoom in places like modals by dividing by the zoom level. This is quite easy by using a CSS property, but does requiring looking through the app for things that don't render correctly.

Notably, this is used in:
- context menu (where you click, but also computing if the context menu should go up/down)
- tooltips (position but also direction)
- modals
- main UI height

Reviewed By: quark-zju

Differential Revision: D51865562

fbshipit-source-id: 787e28c6486b17ae1741d5a6ebab33960c16af94