RebaseOperation: consider in-progress state in optimisticDag
Summary:
It's possible that a partially rebased dag is sent to
RebaseOperation.optimisticDag, this diff attempts to handle it.
Note the uncommitted files list and "head" position were not recorded,
so we cannot reconstruct them during rebase. It might be a good idea
to make operations capture a `dag` so they can reason about what was
the state (ex. head location) when the operation was confirmed.

A easier/cleaner fix is probably to delay all state refreshes until all
operations are completed.

Reviewed By: evangrayk

Differential Revision: D51908788

fbshipit-source-id: fa4b403921d67f49dea098752fd18d21d5350df1