better error message when eden fails to start flock
Summary:
Sometimes when there are two EdenFS's attempting to start at the same time
we run into flock error. It's possible we need to take the lock earlier to
avoid two eden's from getting this far in the process. Seems like this
locking code throws instead of returning false, so maybe this error should just
be swallowed and we fall back to the "another eden is running code path".

While we sort that out let's at least give the user a more informative error
message.

Reviewed By: genevievehelsel

Differential Revision: D51903606

fbshipit-source-id: 062921a6899a9e62b7770a199d329bf233f1168e