Report backingstore retry missing to scuba
Summary:
To get a full picture of the kinds of misses we have at each step adding in the backingstore misses on retry.

NOTE: we could go in a refactor HgImporter to return folly::Try<[Object]Ptr>, clean up the code a bit and unify the retry reporting here, but we are planning to remove that code soon.

Reviewed By: genevievehelsel

Differential Revision: D51784901

fbshipit-source-id: b0cb876c892ba2de077d38df49e73fa4785069b1