repo: get rid of "&mut self" for invalidate_all()
Summary:
Everything uses interior mutability, so `&mut` shouldn't be required.

I tweaked invalidate_metalog to reload the metalog rather than simply clear it out. I think this is more in line with the design where any places holding on to a reference to the metalog will also see the update (through the RwLock).

Reviewed By: quark-zju

Differential Revision: D51966347

fbshipit-source-id: 0433da636d0b6472dcbf36f069bf7bd7c169b8c9