Support CLI and URI lands
Summary: Some lands in some repos don't use the normal land modal. They instead make you go to a URI or run a CLI command. I'm not really sure how common these are, but we support them in the old ISL for ULE lands, so we should support it here.

Reviewed By: quark-zju

Differential Revision: D51992222

fbshipit-source-id: 791f9d98aa9371452b522e5b6be26c4cef585537