Increase size of pages of changed files
Summary:
25 files per page is quite small. Since we support scrolling the list of files, let's increase this limit by a lot. The UI seems smooth up until 500 files per page or so.

Importantly, we also need to make sure our generated file path caches are big enough to fit all these files or it'll thrash and cause bad perf.

With 500 files per page, we only want to fetch 1000 generated files in a batch. Beyond 1000, we warn that the generated files ordering may be incorrect.

If we see this causing perf problems in the wild, we could push this back down or make it configurable. For now I think it works well enough to allow a larger limit by default.

Reviewed By: zzl0

Differential Revision: D52004332

fbshipit-source-id: 648cff36a3e9f51918e0804073db226a8ccd32a7