ssl: add phabricator diff version to ssl display
Summary:
Replace `(local changes)` with an indicator of the diff version that a commit corresponds to.  This gives better feedback when looking at modified diff versions, or older commits.

If the hash of the commit corresponds exactly to a phabricator diff version, we show that diff version after the diff info.  If this is the latest version of the diff, we show `(latest)`.

If the hash doesn't correspond exactly to a version, but one of the commit's predecessors does, we show that version with `+ local changes`.

Reviewed By: quark-zju

Differential Revision: D48682952

fbshipit-source-id: d42f1ccd26d060d25ce1311ef3bd2b060b0576d2