Use commit graph to get unsynced ancestors
Summary:
The current implementation of the `initial-import` command uses the `find_toposorted_unsynced_ancestors` function to get the latest synced ancestors. This function is not efficient, as it traverse the ancestors one by one to get the unsynced ones.

This is especially bad for the `initial-import` command, which is expected to sync everything.

Normally I would update `find_toposorted_unsynced_ancestors` to use the new commit graph for faster traversal, but because there are multiple places using it and it's production critical, we decided offline that it would be best to implement a new function just for the `initial-import` command.

I also added a check to make sure that (a) only one sync config version was used to sync the ancestors and (b) this version is the one provided manually.

If someone **really knows what they're doing**, they can comment out those assertions for local runs.

Reviewed By: mitrandir77

Differential Revision: D52031387

fbshipit-source-id: 2fba416319419e27ac2542e1d52ef88bcb32cfd4