Move Packfile tests into a separate TARGET
Summary:
Currently the packfile implementation and unit tests are part of the same crate. Usually this structure is fine, however in this case it leads to the crate depending on git_types crate specifically only for tests related to DeltaInstructions. Later in the stack, we want to allow git_types to depend on the packfile crate. With the current structure, this is not possible and hence this diff refactors the tests into a TARGET of its own.

The diff also includes a TODO indicating where I need to generate and upload packfile items during the process of GitImport and Remote GitImport

Reviewed By: gustavoavena

Differential Revision: D52070457

fbshipit-source-id: d3adeefe6bd4255b39e3e33d01e122c6e8736c43