Introduce GitPackfileBaseItem as interim packfile state
Summary:
Currently, when importing Git commits, trees, tags or blobs, we just store the raw Git object in our data stores. These objects are then used when generating packfiles/bundles for Git protocol or for landing through Git workflow. When Mark was profiling the bundle generation code, he realized that a major bottleneck is the ZLib encoding that we need to perform on-demand during bundle generation. One way to mitigate this problem is to store the encoded data during the initial object upload, which adds a minimal overhead during write (gitimport, remote-gitimport) but would provide a considerable benefit during read (which is the most prominent workflow in Git).

This diff in particular introduces the GitPackfileBaseItem type which would represent the encoded data that we can directly use in Packfiles

Reviewed By: markbt

Differential Revision: D52073664

fbshipit-source-id: 0cbaf1e52ce29b49bc91269a0b8b1dc2dfd24260