SplitStackEditPanel: show file type changes
Summary:
In case a file has a mode change (ex. regular -> executable), show it as a hint.

I thought about making the mode change editable. But that might
introduce complexities with "split in the middle of the stack" case,
similar to why we disabled text editing for split. So the message is read-only.

Reviewed By: evangrayk

Differential Revision: D52079335

fbshipit-source-id: aa941e37b5bebf196f4e1288e9ff12ea3e1d8cf2