SplitDiffFileHeader: show copyFrom path
Summary:
Show the copyFrom path in the file header. Unlike other places where the full
"copyFrom" path is shown, the common prefix is calculated as an attempt to make
the title short, easier to read.

This is inspired from git CLI,
  $ git commit -m rename
  ...
  rename eden/scm/edenscm/utils/{cborutil.py => c.py} (97%)

I dropped `{ }` because the color difference seems to make the situation obvious enough.

Note: GitHub currently shows a => b in full path.

Reviewed By: evangrayk

Differential Revision: D52087430

fbshipit-source-id: 1d4f2a9e026018c6358c0ffc23d962a41349e6e8