commitStackState: keep copyFrom data in denseSubStack
Summary:
Now we can build file stacks with copyFrom ignored, let's use that in
denseSubStack to preserve copyFrom.

One of the reasons to avoid following copyFrom when building file stacks is
because of tricky cases like forked copy, like:

  commit 1 commit 2
  1.txt -> 2.txt (copied)
       \
        -> 3.txt (also copied)

When moving changes from 1.txt to the "right", it's unclear which one to use.
When moving changes from 2.txt to 1.txt, it's also unclear whether that should
also affect 3.txt (and if 3.txt made some other changes, would it cause
conflicts). By not following renames in file stacks, the above case becomes:

  commit 1 commit 2
  1.txt -> 1.txt
           2.txt
           3.txt

So there is no troublesome "move left/right" between 2.txt to 1.txt.
We can still render the diff of 2.txt against 1.txt.

Reviewed By: evangrayk

Differential Revision: D52087431

fbshipit-source-id: e627caf759b5eb978fe35eefd4f3ac23a96ac79e