FileStackEditorLines: support moving deleted lines with copyFrom right
Summary:
For a `copyFrom` file, we force `aRev` / `aLines` to be the `copyFrom` version
when rendering the diff. However, the `fileStack` does not actually have the
content of `copyFrom`. So deleted (from `copyFrom`) lines won't even exist
in the `fileStack` to be movable via `fileStack.mapAllLines`.

To fix this issue, we detect the situation, patch the `fileStack` temporarily
to complete the move, then unpatch it. Since we know it's moving right, the
content of `aRev` is not supposed to change, so patching and unpatching it
should be fine.

Reviewed By: evangrayk

Differential Revision: D52104331

fbshipit-source-id: 9dfdd510a7e82d3116b150baf20bd122eb484cc9