SplitDiffView: update file header to match split UI
Summary:
D52087430 made split UI show copyFrom path in the file header, which is
similar to GitHub but it also de-dups common prefix to make human parsing
easier.

Update the ComparisonView to match the UI. The copyFrom is changed from a
separated "was renamed from" message to part of the header.

Since the header also has an icon to show file add/remove. I removed the "file
was added/removed" text message to make it a bit more compact.

Reviewed By: evangrayk

Differential Revision: D52106009

fbshipit-source-id: 4bff060691f04c1077e6e577fefc81b1971521bb