SplitDiffFileHeader: define a dedicated type for icons in the header
Summary:
Practically, we only use add/remove/modify icons. Let's define the type
explicitly.

This is partially because the "copied", "renamed" are somewhat ambiguous and
requires extra complexity to figure out. For example, if `a` is copied to both
`b` and `c`, then `a` gets deleted. One of `b` and `c` will show as "renamed"
and the other one as "copied". All copied and renamed files should be newly
created and can fit the "added" icon.

Differential Revision: D52127741

Reviewed By: evangrayk

fbshipit-source-id: 9043daaee8d80c7babead613aae130bf1fa424c8