commitStackState: drop heuristics that turns empty files to deletion
Summary:
Now the split UI can handle "absent" flag properly. There is no need for the
heuristics.

For example, suppose a file has 2 lines and deleted:

  [-] file
  -Line 1
  -Line 2

The user can move a line right:

  [M] file  [-] file
  -Line 1
            -Line 2

Then move the line back to the left:

  [M] file  [-] file
  -Line 1
  -Line 2

Now the first commit makes the file empty and the second deletes the empty
file.

If the empty file state is not desirable, use "move entire file left" on the
second commit, then it will become the initial state.

Differential Revision: D52122438

Reviewed By: evangrayk

fbshipit-source-id: b3af6593771a531f1c102bf1ba999fcb3be3f3a1