SplitDiffFileHeader: tweak `copyFrom` rendering
Summary:
Move the `copyFrom` part to the right side so the actual file path might be a
bit easier to read from left to right. With this change we need `{ }` grouping
to show what part is the common prefix.

Inspired by VS Code, which uses `file (dir) <- fullCopyFromPath` format to
show renames.
 {F1173837201}

Differential Revision: D52110527

Reviewed By: evangrayk

fbshipit-source-id: 932d80a93bd5be84b9f9040a235767f5fd4e30c9