Fix issue where valid, empty blobs are treated as errors
Summary:
When migrating cxx, code needed to be written to determine when a blob fetch was an error. The code was written with the assumption that an empty blob was an error. This is not the case - there are valid, empty blobs. Let's fix that by instead using the absense or presense of an error message as the differentiator.

Also, there was an apparent use after free condition that required the resolve method to ensure that the error string was only used during its lifetime. This was likely true, but not certain. Addressed by converting to a string rather than using char*.

Reviewed By: xavierd

Differential Revision: D52126320

fbshipit-source-id: e9c36c1cf28de211db25271d600cd6c0129801c7