sl: add file change summary
Summary:
This adds a new template function: `pathsummary`.  This function takes a list
of files that have been modified, and summarises them into a limited number of
lines.

Where too many files have been changed in a single directory, the directory is
summarised as `path/to/dir (N files)`, where N is the number of files changed.

Where too many files have been changed in a whole directory tree, the tree is
grouped together and the top-level directory is summarised as
`path/to/top/dir/… (N files)`.

This function is then used to summarise the modifications, additions and
deletions in the current commit in smartlog.  With `--verbose`, all draft
commits are shown.

Reviewed By: muirdm

Differential Revision: D38865981

fbshipit-source-id: b05ce4dff65608ca4d121bc32f83d4ff6527034f