Method for uploading packfile base item in gitimport and remote-gitimport
Summary: Now that we have implemented store and load logic for GitPackfileBaseItem, this diff adds method in the `Uploader` trait used by both gitimport and remote-gitimport to allow storing the packfile base items for raw git objects while they are getting imported. The uploader method for remote-gitimport is just a stub at this point since it requires a corresponding SCS method to allow for that upload, which will be added later in the stack.

Reviewed By: mitrandir77

Differential Revision: D52121827

fbshipit-source-id: 240f24bbb7a2d786c8d15e97d0c54cdfe14425ac