commit_graph: implement minimize_frontier method
Summary:
Implements a method to minimize changeset frontiers by removing changesets that are ancestors of other changesets in the frontier. The set of ancestors of the minimized frontier should be equal to the set of ancestors of the original frontier.

Will be used in the next diff in the `mononoke_newadmin derived-data slice` command to minimize the size of slice boundaries

Reviewed By: mitrandir77

Differential Revision: D52018285

fbshipit-source-id: ac7fc59517e2f028bb7233b989da4d38ff12e919