commit_graph: fix bug in ancestors_difference_segments that causes returning locations for changesets in common
Summary:
Fix a bug in `ancestors_difference_segments` that causes returning locations for changesets that are part of common. This is done by moving the calculation of additional locations in `disjoint_segments` to the part of the code where the head changesets are added to the frontier. These are the cases to consider:
- The frontier doesn't contain the given head changeset. We add a location for it relative to itself (i.e. head~0).
- The frontier contains the given head changeset with an origin that's a head. We add a location relative to the origin head.
- The frontier contains the given head changeset with an origin that's a common changeset. We don't add a location for it because it's part of common.

Afterwards when adding the common changesets of the same generation we should remove any locations that corresponds to any of them. And after that all remaining changesets have a lower generation and so will not affect the locations we've calculated thus far.

Reviewed By: markbt

Differential Revision: D52161500

fbshipit-source-id: e8525e83e102248a1964db75c26ebaad3dd2a36e