commit_graph: implement alternative to hash_to_location
Summary: Implements an alternative to `hash_to_location` that uses the skew binary commit graph instead of the segmented changelog. CommitGraph::changeset_ids_to_locations takes two lists of changesets `heads` and `targets`, and returns a location for all changesets in `targets` that are ancestors of any changeset in `heads`. Each location is of the form `{ cs_id, distance }` which points to the `distance`th ancestor of `cs_id`. Changesets that are acceptable to use in locations are `heads` any parent of any merge ancestor of `heads`.

Reviewed By: markbt

Differential Revision: D52153233

fbshipit-source-id: 607e8140bba1f0beb1f95ff74210af65f5ee5e41