commit_graph: verify that segment parent locations are relative to heads of subsequent segments in unit tests
Summary:
Locations of parents of segments returned from `ancestors_difference_segments` must be relative to heads of subsequently returned segments otherwise they will not be meaningful to the caller. This diff adds the missing verification for this.

After adding this I found that the `test-newadmin-commit-graph-segments.t` integration test started failing because one of the returned locations is relative to a changeset in common. This is because of a wrong assumption when creating the locations map in `disjoint_segments` where we check in `common_skew_ancestors_set` to see if a changeset is an ancestor of common. This is incorrect because this set only contains ancestors reachable through skew binary ancestor edges only (which is useful only in the specific case where we check if for a changeset that has a generation greater than or equal to all of common can follow its skew binary ancestor edge without hitting a common changeset). This will be fixed in the next diff.

Reviewed By: markbt

Differential Revision: D52153236

fbshipit-source-id: 278531263a5c15bda4631cab563611d770fe1cf0