allow forcing mapping change during forwards syncing
Summary:
This diff introduces a way to execute a mapping version
change while syncing a single commit from small->large. See the test changes
to understand how it can be used for repo influx into megarepo.

Compared with the prodecure we used for previous repo merges we don't have to
lock the repositories for this one. This should **just work**.

NOTE: Stack context: I want to show the e2e flow for upcoming repo merge and
test the interaction of forwardsyncer, globalrevs and repos imported from git.
This stack contains the tests showing problems and gradually solves them.

Reviewed By: markbt

Differential Revision: D52120089

fbshipit-source-id: d1f189e86324b1bbb1d2756cb30a19e0f543a911