remove version selection logic from forward syncer
Summary:
This code has 2 problems that make it all wrong:
 * it's basically duplicating what we do during normal sync, so it's not really
   adding much
 * it's using the version used for syncing parents as the version that should
   be used, this is incorrect if there were any version change in the meantime
   and can cause version revert.

Since this logic basically duplicates the logic used on other layers in the
stack, we can get rid of it. This doesn't fix the second problem but at least
reduces the complexity.

NOTE: Stack context: I want to show the e2e flow for upcoming repo merge and
test the interaction of forwardsyncer, globalrevs and repos imported from git.
This stack contains the tests showing problems and gradually solves them.

Reviewed By: markbt

Differential Revision: D52120093

fbshipit-source-id: 212e1e38d225c46c9c27fac359c4eb02a44a0072