test forward syncer interaction with globalrevs
Summary:
Now, that we have globalrev enabled-repo connected to our moonorepo we need to
be careful to not break the globalrev assignment when adding more repos to the
complex. This diff shows that current sync doesn't assign globalrevs on
pushrebase.

NOTE: Stack context: I want to show the e2e flow for upcoming repo merge and
test the interaction of forwardsyncer, globalrevs and repos imported from git.
This stack contains the tests showing problems and gradually solves them.

Reviewed By: gustavoavena

Differential Revision: D52154344

fbshipit-source-id: 8f1e8bb4fdefe6102273591bcb1d084d058957a7