extract get_pushrebase_hooks to separate crate
Summary:
To avoid dependency cycle I need to extract this function to separate crate.
This will allow all pushrebase users to depend on the same way of detemining
pushrebase hooks that should be ran.

This is especially important because pushrebase hooks are ran not only for
validation but also for adding extra data to our DBs like globalrevs and git
mappings.

NOTE: Stack context: I want to show the e2e flow for upcoming repo merge and
test the interaction of forwardsyncer, globalrevs and repos imported from git.
This stack contains the tests showing problems and gradually solves them.

Reviewed By: gustavoavena

Differential Revision: D52154340

fbshipit-source-id: 7d93815d75e23795bc5bffbb878f1792a455efbd