Repository: avoid refreshing states during the first 10 seconds of a running operation
Summary:
See also D51908788. Basically, it's very hard to implement previews for
indeterminate states (status, dag, `.`) when a command is running, because:
- It's impossible for the command to update all the states atomically, so ISL
  can observe a "consistent" state. This is because of the lack of a
  transnational filesystem, and `sl` / `hg` read operations (like `log` and
  `status`) are (mostly) filesystem-lock-free.
- Even if the above problem is solved, `hg log` and `hg status` are 2 separate
  commands run by ISL, so there might be races where the status and log do not
  match perfectly.

Practically, commands like `rebase` might update the commit graph before being
able to move `.`, so `log` and `status` might see an inconsistent state where
the `.` is at the old location keeping a bunch of obsoleted commits. The `status`
can be messy when a non-edenfs checkout is running (on EdenFS it will be an error).

Therefore, it's better to treat indeterminate states as "messy" during a
command and hide them from the users. Practically, the optimistic state fills
the gap and makes the UI feel responsive.

To avoid "stale" UI, the delay is limited to 10 seconds per command.

In the future we might want to revisit areas like:
- Should we use locking to ensure `.`, `log` and `status` are always
  "consistent" when being read out?
- Should we use a single command / operation to read out both `log` and
  `status` without races?
- We probably want `status` to error out during a checkout like edenfs,
  to avoid showing lots of unrelated files.

Reviewed By: evangrayk, zzl0

Differential Revision: D52150066

fbshipit-source-id: d2e3a1d49d389fd2470eb0112c4f7d90659694c8