make extension lists in doctor configurable
Summary:
It's a pain to add and remove extensions from our warn and blocklists in
eden doctor. It it takes a while for things to rollout since the change
is dependent on the eden package which only rolls out once a week.

Let's move these to eden configs so that we can change them with chef and
the new configerator configs.

Reviewed By: genevievehelsel

Differential Revision: D52140236

fbshipit-source-id: 737595f4fdc964b334cc3479a3a4a6369251306d