Replace remotefilelog_file_history_limit tunable with constant
Summary: The tunable `remotefilelog_file_history_limit` was introduced 3.5 years back to avoid pulling in way too large histories. The tunable was added with a value of 1000 and since then has remained the same. IMO this value is very unlikely to change hence it would make sense to have it as a constant instead of a JustKnob or Config value.

Differential Revision: D52200013

fbshipit-source-id: cd7581cd132bafcc55d6b0ca4c7c16b09dd4d060