Replace rollout tunable run_hooks_on_additional_changesets with JK
Summary: The tunable `run_hooks_on_additional_changesets` was added in D23287998 as a rollout tunable so that it can be used as a killswitch in case things go wrong. The relevant chunk of code has been in production for multiple years so it should be safe to remove however a bunch of integration tests rely on this value being false which is why I am replacing it with a JK to prevent breaking things.

Reviewed By: mitrandir77, YousefSalama

Differential Revision: D52200666

fbshipit-source-id: 2e574c415f31c43152d4fa0757860f79b0c56395