Replace hooks_additional_changesets_limit tunable with constant
Summary: The tunable `hooks_additional_changesets_limit` controls the maximum number of changesets that we run the hooks on as a result of bookmark movement. This tunable was created with value 2_00_000 and has remained the same for the past 3.5 years. Its safe to replace it with a constant since it is unlikely to change in the future.

Reviewed By: clara-9

Differential Revision: D52201033

fbshipit-source-id: 56be38df8adb3e3b415ee49bef422256cd09f5e1